etc: GridMain move patterns
